Changes to README.md and XMLParserTests.pck Originally submitted in pull request https://github.com/pbella/Cuis-Ports/pull/2 by hhzl:
Instead of
self assert: error = 'this collection is empty'.

it is necessary to do
self assert: error messageText = 'this collection is empty'.

The README.md now references the more comprehensive ports list at
https://github.com/jvuletich/Cuis/blob/master/ListOfCuisPackages.md
and only lists the ones which are actually in this repository.

(Note: a 3rd 'self assert: error...' scenario was found and corrected as well)